+++ test/src/index.ts
@@ -1,10 +1,12 @@
+import './runner';
+import * as _attrTests from './test-attr';
 import * as _baseTests from './test-base';
-import * as _hubTests from './test-hub';
-import * as _eventTests from './test-event';
+import * as _cssTests from './test-css';
 import * as _domTests from './test-dom';
 import * as _dxTests from './test-dx';
+import * as _eventTests from './test-event';
+import * as _hubTests from './test-hub';
 import * as _utilsTests from './test-utils';
-import * as _attrTests from './test-attr';
 
 export const baseTests = _baseTests;
 export const hubTests = _hubTests;
@@ -13,5 +15,6 @@ export const domTests = _domTests;
 export const dxTests = _dxTests;
 export const utilsTests = _utilsTests;
 export const attrTests = _attrTests;
+export const cssTests = _cssTests;
 
-import './runner';
+