+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 > Legend: `.` minor; `+` addition; `^` enhancement, `*` major; `-` fix; `!` change; 
 
 
+### v0.9.12 PENDING
+
+- `+` adoptStyleSheet - adopt a `CSSObject` to a shadowRoot, using constructable stylesheets if supported or append style element if not.
+- `!^` css - now css returns a `CSSObject` which is composable and can be used in `adoptStyleSheet`
+- `-` @onHub - fix issue where @onHub was not unbound
+- `!` c-base - remove attributeChangedCallback empty implementation (perf impact for little to no value)
+
 ### v0.9.11 Sep 9, 2020
 
 - `+` css - added `css(string)` and css\` tag template to create `<style>` element